+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlayground.swift
@@ -64,7 +64,9 @@ struct PaymentSheetTestPlayground: View {
                         }
                         SettingView(setting: $playgroundController.settings.mode)
                         SettingPickerView(setting: $playgroundController.settings.integrationType)
-                        SettingView(setting: $playgroundController.settings.customerMode)
+                        SettingView(setting: customerModeBinding)
+                        TextField("CustomerId", text: customerIdBinding)
+                            .disabled(true)
                         SettingPickerView(setting: $playgroundController.settings.currency)
                         SettingPickerView(setting: $playgroundController.settings.merchantCountryCode)
                         SettingView(setting: $playgroundController.settings.apmsEnabled)
@@ -113,7 +115,28 @@ struct PaymentSheetTestPlayground: View {
         } set: { newString in
             playgroundController.settings.customCtaLabel = (newString != "") ? newString : nil
         }
-
+    }
+    var customerModeBinding: Binding<PaymentSheetTestPlaygroundSettings.CustomerMode> {
+        Binding<PaymentSheetTestPlaygroundSettings.CustomerMode> {
+            return playgroundController.settings.customerMode
+        } set: { newMode in
+            playgroundController.settings.customerId = nil
+            playgroundController.settings.customerMode = newMode
+        }
+    }
+    var customerIdBinding: Binding<String> {
+        Binding<String> {
+            switch playgroundController.settings.customerMode {
+            case .guest:
+                return ""
+            case .new:
+                return playgroundController.settings.customerId ?? ""
+            case .returning:
+                return playgroundController.settings.customerId ?? ""
+            }
+        } set: { newString in
+            playgroundController.settings.customerId = (newString != "") ? newString : nil
+        }
     }
 }