+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentIntents/STPPaymentIntent.swift
@@ -129,26 +129,8 @@ public class STPPaymentIntent: NSObject {
     /// The optionally expanded PaymentMethod used in this PaymentIntent.
     @objc public let paymentMethod: STPPaymentMethod?
 
-    /// The ordered payment method preference for this PaymentIntent
-    @_spi(STP) public let orderedPaymentMethodTypes: [STPPaymentMethodType]
-
-    /// A list of payment method types that are not activated in live mode, but activated in test mode
-    @_spi(STP) public let unactivatedPaymentMethodTypes: [STPPaymentMethodType]
-
     /// Payment-method-specific configuration for this PaymentIntent.
     @_spi(STP) public let paymentMethodOptions: STPPaymentMethodOptions?
-
-    /// Link-specific settings for this PaymentIntent.
-    @_spi(STP) public let linkSettings: LinkSettings?
-
-    /// Country code of the user.
-    @_spi(STP) public let countryCode: String?
-
-    /// Country code of the merchant.
-    @_spi(STP) public let merchantCountryCode: String?
-
-    /// Card brand choice settings for the merchant.
-    @_spi(STP) public let cardBrandChoice: STPCardBrandChoice?
 
     /// :nodoc:
     @objc public override var description: String {
@@ -163,15 +145,12 @@ public class STPPaymentIntent: NSObject {
             "captureMethod = \(String(describing: allResponseFields["capture_method"] as? String))",
             "clientSecret = <redacted>",
             "confirmationMethod = \(String(describing: allResponseFields["confirmation_method"] as? String))",
-            "countryCode = \(String(describing: countryCode))",
             "created = \(created)",
             "currency = \(currency)",
             "description = \(String(describing: stripeDescription))",
             "lastPaymentError = \(String(describing: lastPaymentError))",
-            "linkSettings = \(String(describing: linkSettings))",
             "livemode = \(livemode)",
             "nextAction = \(String(describing: nextAction))",
-            "merchantCountryCode = \(String(describing: merchantCountryCode))",
             "paymentMethodId = \(String(describing: paymentMethodId))",
             "paymentMethod = \(String(describing: paymentMethod))",
             "paymentMethodOptions = \(String(describing: paymentMethodOptions))",
@@ -181,8 +160,6 @@ public class STPPaymentIntent: NSObject {
             "shipping = \(String(describing: shipping))",
             "sourceId = \(String(describing: sourceId))",
             "status = \(String(describing: allResponseFields["status"] as? String))",
-            "unactivatedPaymentMethodTypes = \(allResponseFields.stp_array(forKey: "unactivated_payment_method_types") ?? [])",
-            "cardBrandChoice = \(String(describing: cardBrandChoice))",
         ]
 
         return "<\(props.joined(separator: "; "))>"
@@ -195,15 +172,11 @@ public class STPPaymentIntent: NSObject {
         captureMethod: STPPaymentIntentCaptureMethod,
         clientSecret: String,
         confirmationMethod: STPPaymentIntentConfirmationMethod,
-        countryCode: String?,
         created: Date,
         currency: String,
         lastPaymentError: STPPaymentIntentLastPaymentError?,
-        linkSettings: LinkSettings?,
         livemode: Bool,
         nextAction: STPIntentAction?,
-        merchantCountryCode: String?,
-        orderedPaymentMethodTypes: [STPPaymentMethodType],
         paymentMethod: STPPaymentMethod?,
         paymentMethodId: String?,
         paymentMethodOptions: STPPaymentMethodOptions?,
@@ -214,9 +187,7 @@ public class STPPaymentIntent: NSObject {
         sourceId: String?,
         status: STPPaymentIntentStatus,
         stripeDescription: String?,
-        stripeId: String,
-        unactivatedPaymentMethodTypes: [STPPaymentMethodType],
-        cardBrandChoice: STPCardBrandChoice?
+        stripeId: String
     ) {
         self.allResponseFields = allResponseFields
         self.amount = amount
@@ -224,15 +195,11 @@ public class STPPaymentIntent: NSObject {
         self.captureMethod = captureMethod
         self.clientSecret = clientSecret
         self.confirmationMethod = confirmationMethod
-        self.countryCode = countryCode
         self.created = created
         self.currency = currency
         self.lastPaymentError = lastPaymentError
-        self.linkSettings = linkSettings
         self.livemode = livemode
         self.nextAction = nextAction
-        self.merchantCountryCode = merchantCountryCode
-        self.orderedPaymentMethodTypes = orderedPaymentMethodTypes
         self.paymentMethod = paymentMethod
         self.paymentMethodId = paymentMethodId
         self.paymentMethodOptions = paymentMethodOptions
@@ -244,8 +211,6 @@ public class STPPaymentIntent: NSObject {
         self.status = status
         self.stripeDescription = stripeDescription
         self.stripeId = stripeId
-        self.unactivatedPaymentMethodTypes = unactivatedPaymentMethodTypes
-        self.cardBrandChoice = cardBrandChoice
         super.init()
     }
 }
@@ -255,32 +220,6 @@ extension STPPaymentIntent: STPAPIResponseDecodable {
 
     @objc
     public class func decodedObject(fromAPIResponse response: [AnyHashable: Any]?) -> Self? {
-        guard let response = response else {
-            return nil
-        }
-
-        if let paymentMethodPrefDict = response["payment_method_preference"] as? [AnyHashable: Any],
-            let paymentIntentDict = paymentMethodPrefDict["payment_intent"] as? [AnyHashable: Any],
-            let orderedPaymentMethodTypes = paymentMethodPrefDict["ordered_payment_method_types"]
-                as? [String]
-        {
-            // Consolidates expanded payment_intent and ordered_payment_method_types into singular dict for decoding
-            var dict = paymentIntentDict
-            dict["country_code"] = paymentMethodPrefDict["country_code"]
-            dict["merchant_country"] = response["merchant_country"]
-            dict["ordered_payment_method_types"] = orderedPaymentMethodTypes
-            dict["unactivated_payment_method_types"] = response["unactivated_payment_method_types"]
-            dict["link_settings"] = response["link_settings"]
-            dict["payment_method_specs"] = response["payment_method_specs"]
-            dict["card_brand_choice"] = response["card_brand_choice"]
-            return decodeSTPPaymentIntentObject(fromAPIResponse: dict)
-        } else {
-            return decodeSTPPaymentIntentObject(fromAPIResponse: response)
-        }
-    }
-
-    class func decodeSTPPaymentIntentObject(fromAPIResponse response: [AnyHashable: Any]?) -> Self?
-    {
         guard let dict = response,
             let stripeId = dict["id"] as? String,
             let clientSecret = dict["client_secret"] as? String,
@@ -299,9 +238,6 @@ extension STPPaymentIntent: STPAPIResponseDecodable {
         )
         let setupFutureUsageString = dict["setup_future_usage"] as? String
         let canceledAtUnixTime = dict["canceled_at"] as? TimeInterval
-        let unactivatedPaymentTypes = STPPaymentMethod.paymentMethodTypes(
-            from: dict["unactivated_payment_method_types"] as? [String] ?? []
-        )
         return STPPaymentIntent(
             allResponseFields: dict,
             amount: amount,
@@ -314,22 +250,14 @@ extension STPPaymentIntent: STPAPIResponseDecodable {
             confirmationMethod: STPPaymentIntentConfirmationMethod.confirmationMethod(
                 from: dict["confirmation_method"] as? String ?? ""
             ),
-            countryCode: dict["country_code"] as? String,
             created: Date(timeIntervalSince1970: createdUnixTime),
             currency: currency,
             lastPaymentError: STPPaymentIntentLastPaymentError.decodedObject(
                 fromAPIResponse: dict["last_payment_error"] as? [AnyHashable: Any]
             ),
-            linkSettings: LinkSettings.decodedObject(
-                fromAPIResponse: dict["link_settings"] as? [AnyHashable: Any]
-            ),
             livemode: livemode,
             nextAction: STPIntentAction.decodedObject(
                 fromAPIResponse: dict["next_action"] as? [AnyHashable: Any]
-            ),
-            merchantCountryCode: dict["merchant_country"] as? String,
-            orderedPaymentMethodTypes: STPPaymentMethod.paymentMethodTypes(
-                from: dict["ordered_payment_method_types"] as? [String] ?? paymentMethodTypeStrings
             ),
             paymentMethod: paymentMethod,
             paymentMethodId: paymentMethod?.stripeId ?? dict["payment_method"] as? String,
@@ -346,9 +274,7 @@ extension STPPaymentIntent: STPAPIResponseDecodable {
             sourceId: dict["source"] as? String,
             status: STPPaymentIntentStatus.status(from: rawStatus),
             stripeDescription: dict["description"] as? String,
-            stripeId: stripeId,
-            unactivatedPaymentMethodTypes: unactivatedPaymentTypes,
-            cardBrandChoice: STPCardBrandChoice.decodedObject(fromAPIResponse: dict["card_brand_choice"] as? [AnyHashable: Any])
+            stripeId: stripeId
         ) as? Self
     }
 }