+++ StripeUICore/StripeUICore/Source/Elements/DropdownFieldElement.swift
@@ -263,6 +263,7 @@ extension DropdownFieldElement: PickerFieldViewDelegate {
         if shouldAutoAdvance {
             DispatchQueue.main.async { [weak self] in
                 guard let self = self else { return }
+                self.delegate?.didUpdate(element: self)
                 self.delegate?.continueToNextField(element: self)
             }
         }